Minor changes in constants , backend and pypi deployment

+++ zseeker/constants.py
@@ -7,12 +7,12 @@ class Params:
     AC_weight: float = 1.25
     AT_weight: float = 0.5
     display_sequence_score: int = 0
-    mismatch_penalty_starting_value: int = 3.0 
-    mismatch_penalty_linear_delta: int = 3.0
+    mismatch_penalty_starting_value: int = 3
+    mismatch_penalty_linear_delta: int = 3
     mismatch_penalty_type: str = "linear"
     mismatch_penalty_choices: tuple[str] = ("linear", "exponential")
     method_choices: tuple[str] = ("transitions", "coverage", "layered")
-    consecutive_AT_scoring: tuple[int] = (0.5, 0.5, 0.5, 0.5, 0.0, 0.0, -5.0, -100.0)
+    consecutive_AT_scoring: tuple[float] = (0.5, 0.5, 0.5, 0.5, 0.0, 0.0, -5.0, -100.0)
     n_jobs: int = 8
     output_dir: str = "zdna_extractions"
     cadence_reward: float = 1
